pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "NutriCore_backend"
version = "0.1.0"
description = ""
authors = [
    {name = "k1fir38",email = "k1fir38@yandex.ru"}
]
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.135.3,<0.136.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "asyncio (>=4.0.0,<5.0.0)",
    "pydantic (>=2.13.0,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "greenlet (>=3.4.0,<4.0.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
